# api/docs/swagger.yaml
basePath: /
definitions:
  auth.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  auth.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  expense.CreateExpenseRequest:
    properties:
      amount:
        type: number
      category:
        type: string
      description:
        type: string
    required:
    - amount
    - category
    type: object
  expense.DeleteExpenseResponse:
    properties:
      message:
        type: string
    type: object
  expense.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  expense.ExpenseResponse:
    properties:
      amount:
        type: number
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  expense.UpdateExpenseRequest:
    properties:
      amount:
        type: number
      category:
        type: string
      description:
        type: string
    required:
    - amount
    - category
    type: object
  server.CurrentUserResponse:
    properties:
      email: {}
      user_id: {}
    type: object
  server.HealthResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: This is a sample expense tracker server.
  title: Expense Tracker API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return tokens
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: User Login
      tags:
      - auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Issue a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: User Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: User Registration
      tags:
      - auth
  /api/v1/expenses:
    get:
      description: List all expenses for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/expense.ExpenseResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List expenses
      tags:
      - expenses
    post:
      consumes:
      - application/json
      description: Create a new expense for the authenticated user
      parameters:
      - description: Expense data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/expense.CreateExpenseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/expense.ExpenseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create expense
      tags:
      - expenses
  /api/v1/expenses/{id}:
    delete:
      description: Delete one expense by ID for the authenticated user
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/expense.DeleteExpenseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete expense
      tags:
      - expenses
    get:
      description: Get one expense by ID for the authenticated user
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/expense.ExpenseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get expense
      tags:
      - expenses
    put:
      consumes:
      - application/json
      description: Update one expense by ID for the authenticated user
      parameters:
      - description: Expense ID
        in: path
        name: id
        required: true
        type: integer
      - description: Expense data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/expense.UpdateExpenseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/expense.ExpenseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/expense.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update expense
      tags:
      - expenses
  /api/v1/me:
    get:
      description: Get the authenticated user's token claims
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.CurrentUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Current user
      tags:
      - auth
  /health:
    get:
      description: Check API and database health
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.HealthResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.HealthResponse'
      summary: Health check
      tags:
      - system
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"